refactored private chat and group chat recyclerview adapters,wrapped linear-layout managers, replaced unneccessary listeners with firebase database get() method

+++ app/src/main/java/com/colley/android/view/dialog/IssueOptionsBottomSheetDialogFragment.kt
@@ -9,8 +9,10 @@ import android.view.View.GONE
 import android.view.View.VISIBLE
 import android.view.ViewGroup
 import android.widget.Toast
+import androidx.appcompat.app.AlertDialog
 import com.colley.android.databinding.BottomSheetDialogFragmentIssueOptionsBinding
 import com.colley.android.databinding.BottomSheetDialogFragmentMoreBinding
+import com.firebase.ui.auth.AuthUI
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.auth.FirebaseUser
@@ -82,44 +84,55 @@ class IssueOptionsBottomSheetDialogFragment (
             }
 
             deleteIssue.setOnClickListener {
-                issueId?.let { issueId ->
-                    //delete issue from database, pass a DatabaseReference.onCompletionListener
-                    dbRef.child("issues").child(issueId).setValue(null) { error, ref ->
-                        //if no error, continue further operations
-                        if(error == null) {
-                            //decrease issues count
-                            dbRef.child("issuesCount").runTransaction(
-                                object : Transaction.Handler {
-                                    override fun doTransaction(currentData: MutableData): Transaction.Result {
-                                        var count = currentData.getValue(Int::class.java)
-                                        if (count != null){
-                                            currentData.value = count--
-                                        }
-                                        currentData.value = count
-                                        //set database count value to the new update
-                                        return Transaction.success(currentData)
-                                    }
 
-                                    override fun onComplete(
-                                        error: DatabaseError?,
-                                        committed: Boolean,
-                                        currentData: DataSnapshot?
-                                    ) {}
+                AlertDialog.Builder(parentContext)
+                    .setMessage("Are you sure you want to delete this issue?")
+                    .setPositiveButton("Yes") { dialog, which ->
+                        issueId?.let { issueId ->
+                            //delete issue from database, pass a DatabaseReference.onCompletionListener
+                            dbRef.child("issues").child(issueId).setValue(null) { error, ref ->
+                                //if no error, continue further operations
+                                if(error == null) {
+                                    //decrease issues count
+                                    dbRef.child("issuesCount").runTransaction(
+                                        object : Transaction.Handler {
+                                            override fun doTransaction(currentData: MutableData): Transaction.Result {
+                                                var count = currentData.getValue(Int::class.java)
+                                                if (count != null){
+                                                    currentData.value = count--
+                                                }
+                                                currentData.value = count
+                                                //set database count value to the new update
+                                                return Transaction.success(currentData)
+                                            }
+
+                                            override fun onComplete(
+                                                error: DatabaseError?,
+                                                committed: Boolean,
+                                                currentData: DataSnapshot?
+                                            ) {}
+
+                                        }
+                                    )
+
+                                    //set issue comments to null
+                                    dbRef.child("issue-comments").child(issueId).setValue(null)
+
+                                    //trigger listeners
+                                    moreOptionsDialogListener.onDeleteIssue(issueId)
 
                                 }
-                            )
-
-                            //set issue comments to null
-                            dbRef.child("issue-comments").child(issueId).setValue(null)
-
-                            //trigger listeners
-                            moreOptionsDialogListener.onDeleteIssue(issueId)
+                                this@IssueOptionsBottomSheetDialogFragment.dismiss()
+                            }
 
                         }
-                        this@IssueOptionsBottomSheetDialogFragment.dismiss()
-                    }
+                        //dismiss alert dialog
+                        dialog.dismiss()
 
-                }
+                    }.setNegativeButton("No") { dialog, which ->
+                        dialog.dismiss()
+                    }.show()
+
 
 
             }
